fix: created and separated a new layout for sidebar
