+++ src/common/sidebar/components/buttonBox/style.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+export const ButtonBox = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  padding-bottom: 30px;
+`;
