+++ src/common/sidebar/style.ts
@@ -10,19 +10,3 @@ export const Sidebar = styled.div`
   justify-content: space-between;
   box-shadow: 10px 10px 10px ${Theme.colors.gray2};
 `;
-export const CategoryBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: fit-content;
-  font-size: ${Theme.fontSizes.body4};
-  font-weight: ${Theme.fontWeights.normal};
-`;
-export const ButtonBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  padding-bottom: 30px;
-`;
